move options to runtime
