+++ toolruntime/responses_stream.go
@@ -14,7 +14,6 @@ import (
 	"github.com/modelcontextprotocol/go-sdk/mcp"
 
 	"github.com/tinfoilsh/confidential-model-router/manager"
-	"github.com/tinfoilsh/confidential-model-router/toolcontext"
 	"github.com/tinfoilsh/confidential-model-router/toolruntime/citations"
 )
 
@@ -894,7 +893,7 @@ func runResponsesStreaming(
 	modelName string,
 	requestHeaders http.Header,
 	prompt *mcp.GetPromptResult,
-	routerOpts *toolcontext.RouterOptions,
+	routerOpts *RouterOptions,
 	dl *devLog,
 ) error {
 	flusher, ok := w.(http.Flusher)
